requirements.txt
motor==3.7.1
beanie==2.0.1
fastapi==0.124.4
envyaml==0.1910
uvicorn==0.38.0
aiokafka==0.12.0
python-logstash-async==3.0.0
cryptography==46.0.3
pytest==8.3.4
pytest-asyncio==0.25.1
pytest-mock==3.14.0
ruff==0.1.9